💄 UI : SideSection 컴포넌트 대체 수정
